Add unique keys to components, array and elements
Since no id or uuid is provided, because there is no database. A basic
function to generate unique keys is used. For the array of arrays it
just concatenate all the elements, uniqueness is ensured when the
combinations are generated. The element component just concatenate the
element with its index.